Started expanding tests and adding new parts
