+++ src/config/semrush-client.php
@@ -4,6 +4,7 @@ use League\OAuth2\Client\Provider\Exception\IdentityProviderException;
 use League\OAuth2\Client\Provider\GenericProvider;
 use League\OAuth2\Client\Token\AccessTokenInterface;
 use Yoast\WP\SEO\Exceptions\OAuth\OAuth_Authentication_Failed_Exception;
+use Yoast\WP\SEO\Helpers\Options_Helper;
 use Yoast\WP\SEO\Routes\SEMrush_Route;
 
 /**
@@ -17,6 +18,11 @@ class SEMrush_Client {
 	 */
 	private $provider;
 
+	/**
+	 * @var SEMrush_Token_Manager
+	 */
+	private $token_manager;
+
 	public function __construct() {
 		$this->provider = new GenericProvider( [
 			'clientId'                => '',
@@ -29,18 +35,31 @@ class SEMrush_Client {
 	}
 
 	/**
-	 * @param $code
+	 * Gets the access token based on the passed code.
 	 *
-	 * @return \Exception|IdentityProviderException|AccessTokenInterface
+	 * @param string $code The code to send.
+	 *
+	 * @return bool
 	 * @throws OAuth_Authentication_Failed_Exception
+	 * @throws \Yoast\WP\SEO\Exceptions\OAuth\OAuth_Expired_Token_Exception
+	 * @throws \Yoast\WP\SEO\Exceptions\OAuth\OAuth_Failed_Token_Storage_Exception
 	 */
 	public function get_access_tokens( $code ) {
 		try {
-			return $this->provider->getAccessToken( 'authorization_code', [
+			$response = $this->provider->getAccessToken( 'authorization_code', [
 				'code' => $code,
 			] );
+
+			// Store tokens
+			$this->token_manager = new SEMrush_Token_Manager( $response, new Options_Helper() );
+
+			if ( $this->token_manager->store() === true ) {
+
+			}
 		} catch ( IdentityProviderException $e ) {
-			throw new OAuth_Authentication_Failed_Exception( $code, $e );
+			throw new OAuth_Authentication_Failed_Exception( $e );
+		} catch ( OAuth_Expired_Token_Exception $e ) {
+			throw new OAuth_Authentication_Failed_Exception( $e );
 		}
 	}
