+++ src/config/semrush-token-manager.php
@@ -24,19 +24,50 @@ class SEMrush_Token_Manager {
 	/**
 	 * SEMrush_Token_Manager constructor.
 	 *
-	 * @param AccessTokenInterface $response
 	 * @param Options_Helper       $options_helper
-	 *
-	 * @throws OAuth_Expired_Token_Exception
 	 */
-	public function __construct( AccessTokenInterface $response, Options_Helper $options_helper ) {
+	public function __construct( Options_Helper $options_helper ) {
+		$this->options_helper = $options_helper;
+	}
+
+	public function from_response( AccessTokenInterface $response ) {
 		// If, for some reason, the token is already considered expired, throw an exception.
 		if ( $response->hasExpired() === true ) {
 			throw new OAuth_Expired_Token_Exception();
 		}
 
 		$this->response       = $response;
-		$this->options_helper = $options_helper;
+	}
+
+	/**
+	 * Gets the original response.
+	 *
+	 * @return AccessTokenInterface The original response.
+	 */
+	public function get_original_response() {
+		return $this->response;
+	}
+
+	/**
+	 * Saves the tokens to storage.
+	 *
+	 * @return bool Whether or not the tokens were successfully saved.
+	 * @throws OAuth_Failed_Token_Storage_Exception
+	 */
+	public function store() {
+		$saved = $this->options_helper->set( 'yst_semrush_tokens', [
+			'access_token'  => $this->get_access_token(),
+			'refresh_token' => $this->get_refresh_token(),
+			'expires'       => $this->get_expires(),
+			'has_expired'   => $this->has_expired(),
+		] );
+
+		// Something went wrong in the saving process.
+		if ( $saved === null || $saved === false ) {
+			throw new OAuth_Failed_Token_Storage_Exception();
+		}
+
+		return $saved;
 	}
 
 	/**
@@ -76,37 +107,6 @@ class SEMrush_Token_Manager {
 	}
 
 	/**
-	 * Gets the original response.
-	 *
-	 * @return AccessTokenInterface The original response.
-	 */
-	public function get_original_response() {
-		return $this->response;
-	}
-
-	/**
-	 * Saves the tokens to storage.
-	 *
-	 * @return bool Whether or not the tokens were successfully saved.
-	 * @throws OAuth_Failed_Token_Storage_Exception
-	 */
-	public function store() {
-		$saved = $this->options_helper->set( 'yst_semrush_tokens', [
-			'access_token'  => $this->get_access_token(),
-			'refresh_token' => $this->get_refresh_token(),
-			'expires'       => $this->get_expires(),
-			'has_expired'   => $this->has_expired(),
-		] );
-
-		// Something went wrong in the saving process.
-		if ( $saved === null || $saved === false ) {
-			throw new OAuth_Failed_Token_Storage_Exception();
-		}
-
-		return $saved;
-	}
-
-	/**
 	 * Retrieves the tokens from storage.
 	 *
 	 * @return array The tokens array. Returns an empty array if none exist.
@@ -114,7 +114,7 @@ class SEMrush_Token_Manager {
 	public function get_from_storage() {
 		$tokens = $this->options_helper->get( 'yst_semrush_tokens' );
 
-		if ( $tokens === null ) {
+		if ( ! is_array( $tokens ) || $tokens === null ) {
 			return [];
 		}
